requirements.txt
amqp==2.5.2
asgiref==3.2.3
autopep8==1.5.3
beautifulsoup4==4.8.2
billiard==3.6.1.0
celery==4.4.0
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.2
django-allauth==0.42.0
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-cors-headers==3.4.0
django-environ==0.4.5
django-heroku==0.3.1
django-rest-auth==0.9.5
django-timezone-field==4.0
djangorestframework==3.11.0
drf-yasg==1.17.1
gunicorn==20.0.4
httplib2==0.17.0
idna==2.9
importlib-metadata==1.4.0
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
kombu==4.6.7
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==20.4
psycopg2==2.8.5
pycodestyle==2.6.0
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.1
python-http-client==3.2.7
python3-openid==3.1.0
pytz==2019.3
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
sendgrid==6.4.2
six==1.15.0
soupsieve==1.9.5
sqlparse==0.3.0
starkbank-ecdsa==1.0.0
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.9
vine==1.3.0
whitenoise==5.1.0
youtube-dl==2020.6.16.1
zipp==2.1.0
